fix newly downloaded models showing up as the helmet model
when you download a new model, the player using it needs to swap to a different model and back again or else you won't see the new model. That player also has to be in the same visibility area as you, but you don't have to be looking at the player. Big models (bytes) may freeze the client for a fraction of a second during the swap.